Outsource plot_[prior_|posterior_]predictions from numpyro_backend to base
