+++ pymob/inference/numpyro_backend.py
@@ -1098,191 +1098,6 @@ class NumpyroBackend(InferenceBackend):
             fig = plt.gcf()
             fig.savefig(f"{self.simulation.output_path}/pairs_posterior.png")
 
-    def plot_prior_predictions(
-            self, data_variable: str, x_dim: str, ax=None, subset={}, 
-            n=None, seed=None, plot_preds_without_obs=False,
-            prediction_data_variable: Optional[str] = None,
-            **plot_kwargs
-        ):
-        if n is None:
-            n = self.n_predictions
-
-        if seed is None:
-            seed = self.config.simulation.seed
-        
-        idata = self.prior_predictions(
-            n=n, 
-            # seed only controls the parameters samples drawn from posterior
-            seed=seed
-        )
-
-        ax = self.plot_predictions(
-            observations=self.simulation.observations,
-            predictions=idata.prior_predictive, # type: ignore
-            data_variable=data_variable,
-            plot_preds_without_obs=plot_preds_without_obs,
-            x_dim=x_dim,
-            ax=ax,
-            subset=subset,
-            prediction_data_variable=prediction_data_variable,
-            **plot_kwargs,
-        )
-
-        return ax
-
-    def plot_posterior_predictions(
-            self, data_variable: str, x_dim: str, ax=None, subset={},
-            n=None, seed=None, plot_preds_without_obs=False,
-            prediction_data_variable: Optional[str] = None,
-            **plot_kwargs
-        ):
-        # TODO: This method should be trashed. It is not really useful
-        if n is None:
-            n = self.n_predictions
-        
-        if seed is None:
-            seed = self.config.simulation.seed
-        
-        predictions = self.posterior_predictions(
-            n=n, 
-            # seed only controls the parameters samples drawn from posterior
-            seed=seed
-        )
-        
-        ax = self.plot_predictions(
-            observations=self.simulation.observations,
-            predictions=predictions,
-            data_variable=data_variable,
-            plot_preds_without_obs=plot_preds_without_obs,
-            x_dim=x_dim,
-            ax=ax,
-            subset=subset,
-            prediction_data_variable=prediction_data_variable,
-            **plot_kwargs
-        )
-
-        return ax
-
-    def plot(self):
-        self.plot_diagnostics()
-
-        plot = self.config.inference.plot
-        if plot is None:
-            return
-        elif isinstance(plot, str):
-            try:
-                plot_func = getattr(self.simulation, plot)
-                plot_func(self.simulation)
-            except AttributeError:
-                warnings.warn(
-                    f"Plot function {plot} was not found in the plot.py module "+
-                    "Make sure the name has been spelled correctly or try to "+
-                    "set the function directly to 'sim.config.inference.plot'.",
-                    category=UserWarning
-                )
-        else:
-            plot(self.simulation)
-
-    def plot_predictions(
-            self, 
-            observations,
-            predictions,
-            data_variable: str,
-            x_dim: str, 
-            ax=None, 
-            plot_preds_without_obs=False,
-            subset={},
-            mode: Literal["mean+hdi", "draws"]="mean+hdi",
-            plot_options: Dict={"obs": {}, "pred_mean": {}, "pred_draws": {}, "pred_hdi": {}},
-            prediction_data_variable: Optional[str] = None,
-        ):
-        # filter subset coordinates present in data_variable
-        subset = {k: v for k, v in subset.items() if k in observations.coords}
-        
-        if prediction_data_variable is None:
-            prediction_data_variable = data_variable + "_obs"
-
-        # select subset
-        if prediction_data_variable in predictions:
-            preds = predictions.sel(subset)[prediction_data_variable]
-        else:
-            raise KeyError(
-                f"{prediction_data_variable} was not found in the predictions "+
-                "consider specifying the data variable for the predictions "+
-                "explicitly with the option `prediction_data_variable`."
-            )
-        try:
-            obs = observations.sel(subset)[data_variable]
-        except KeyError:
-            obs = preds.copy().mean(dim=("chain", "draw"))
-            obs.values = np.full_like(obs.values, np.nan)
-        
-        # stack all dims that are not in the time dimension
-        if len(obs.dims) == 1:
-            # add a dummy batch dimension
-            obs = obs.expand_dims("batch")
-            obs = obs.assign_coords(batch=[0])
-
-            preds = preds.expand_dims("batch")
-            preds = preds.assign_coords(batch=[0])
-
-
-        stack_dims = [d for d in obs.dims if d not in [x_dim, "chain", "draw"]]
-        obs = obs.stack(i=stack_dims)
-        preds = preds.stack(i=stack_dims)
-        N = len(obs.coords["i"])
-            
-        hdi = az.hdi(preds, .95)[f"{prediction_data_variable}"]
-
-        if ax is None:
-            ax = plt.subplot(111)
-        
-        y_mean = preds.mean(dim=("chain", "draw"))
-
-        for i in obs.i:
-            if obs.sel(i=i).isnull().all() and not plot_preds_without_obs:
-                # skip plotting combinations, where all values are NaN
-                continue
-            
-            if mode == "mean+hdi":
-                kwargs_hdi = dict(color="black", alpha=0.1)
-                kwargs_hdi.update(plot_options.get("pred_hdi", {}))
-                ax.fill_between(
-                    preds[x_dim].values, *hdi.sel(i=i).values.T, # type: ignore
-                    **kwargs_hdi
-                )
-
-                kwargs_mean = dict(color="black", lw=1, alpha=max(1/N, 0.05))
-                kwargs_mean.update(plot_options.get("pred_mean", {}))
-                ax.plot(
-                    preds[x_dim].values, y_mean.sel(i=i).values, 
-                    **kwargs_mean
-                )
-            elif mode == "draws":
-                kwargs_draws = dict(color="black", lw=0.5, alpha=max(1/N, 0.05))
-                kwargs_draws.update(plot_options.get("pred_draws", {}))
-                ys = preds.sel(i=i).stack(sample=("chain", "draw"))
-                ax.plot(
-                    preds[x_dim].values, ys.values, 
-                    **kwargs_draws
-                )
-            else:
-                raise NotImplementedError(
-                    f"Mode '{mode}' not implemented. "+
-                    "Choose 'mean+hdi' or 'draws'."
-                )
-
-            kwargs_obs = dict(marker="o", ls="", ms=3, color="tab:blue")
-            kwargs_obs.update(plot_options.get("obs", {}))
-            ax.plot(
-                obs[x_dim].values, obs.sel(i=i).values, 
-                **kwargs_obs
-            )
-        
-        ax.set_ylabel(data_variable)
-        ax.set_xlabel(x_dim)
-
-        return ax
 
     # This is a separate script!    
     def combine_chains(self, chain_location="chains", drop_extra_vars=[], cluster_deviation="std"):
